requirements.txt
requests
requests_cache
hanziconvert